add poetry
