+++ pile/scripts/run_group_stats.sh
@@ -1,1 +1,2 @@
+#!/bin/bash
 python3 check_percentage_non_c4.py --input_dir
